# application/backend/certiblock/docs/swagger.yaml
definitions:
  controllers.PrivateKey:
    properties:
      private_key:
        type: string
    type: object
  data.BCCertificateInput:
    properties:
      file:
        type: string
      student_public_key:
        type: string
      university_private_key:
        type: string
    type: object
  data.BCCertificateOutput:
    properties:
      C_S:
        type: string
      C_U:
        type: string
      C_X:
        type: string
      S:
        type: string
      S0:
        type: string
      hash:
        type: string
      id:
        type: integer
      student_public_key:
        type: string
      university_private_key:
        type: string
    type: object
  data.BCEnrollmentCertificateOutput:
    properties:
      hash:
        type: string
      id:
        type: integer
      student_public_key:
        type: string
      university_public_key:
        type: string
    type: object
  data.BCStudentInput:
    properties:
      public_key:
        type: string
    type: object
  data.BCStudentOutput:
    properties:
      id:
        type: integer
      public_key:
        type: string
    type: object
  data.BCUniversityInput:
    properties:
      admin_approval_key:
        type: string
      name:
        type: string
      private_key:
        type: string
    type: object
  data.BCUniversityOutput:
    properties:
      id:
        type: integer
      name:
        type: string
      public_key:
        type: string
    type: object
  data.CountryOutput:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  gin.H:
    additionalProperties: {}
    type: object
info:
  contact: {}
paths:
  /api/universities:
    post:
      consumes:
      - application/json
      description: Register a university
      parameters:
      - description: University data
        in: body
        name: university
        required: true
        schema:
          $ref: '#/definitions/data.BCUniversityInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.BCUniversityOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Register a university
      tags:
      - universities
  /api/universities/certificate:
    post:
      consumes:
      - application/json
      description: Issue a certificate
      parameters:
      - description: Certificate data
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/data.BCCertificateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.BCCertificateOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Issue a certificate
      tags:
      - universities
  /api/universities/enroll:
    post:
      consumes:
      - application/json
      description: Enroll a student to a university
      parameters:
      - description: Student public key
        in: body
        name: student_public_key
        required: true
        schema:
          type: string
      - description: CCCD hash
        in: body
        name: hash
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.BCEnrollmentCertificateOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Enroll a student to a university
      tags:
      - universities
  /api/universities/info:
    post:
      consumes:
      - application/json
      description: Get information about a university using its private key
      parameters:
      - description: Private key
        in: body
        name: private_key
        required: true
        schema:
          $ref: '#/definitions/controllers.PrivateKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.BCUniversityOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get university information
      tags:
      - universities
  /countries:
    get:
      description: Get all countries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.CountryOutput'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all countries
      tags:
      - countries
  /countries/{id}:
    get:
      description: Get a country by ID
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.CountryOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a country by ID
      tags:
      - countries
  /students:
    post:
      consumes:
      - application/json
      description: Register a student
      parameters:
      - description: Student data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/data.BCStudentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.BCStudentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Register a student
      tags:
      - students
  /students/{id}:
    get:
      description: Get a student by public key
      parameters:
      - description: Student public key
        in: path
        name: public_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.BCStudentOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a student by public key
      tags:
      - students
swagger: "2.0"